fix: structure adjusts
