+++ src/main.ts
@@ -1,33 +1,9 @@
 import 'reflect-metadata';
-import express from 'express';
-import cors from 'cors';
 import { config } from 'dotenv';
+import app from './app';
 import dataSource from './config/typeorm.config';
-import eventRoutes from './routes/events';
-import participantRoutes from './routes/participants';
 
 config();
-
-const app = express();
-app.use(express.json());
-
-app.use(cors({
-  origin: true,
-  credentials: true
-}));
-
-app.use((req, res, next) => {
-  console.log(`${req.method} ${req.path}`);
-  next();
-});
-
-app.use('/events', eventRoutes);
-app.use('/participants', participantRoutes);
-
-app.use((err: Error, req: express.Request, res: express.Response, next: express.NextFunction) => {
-  console.error('Error:', err);
-  res.status(500).json({ message: 'Internal server error' });
-});
 
 const PORT = process.env.PORT || 3000;
