+++ src/modules/participant/controllers/participant.controller.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import dataSource from '../config/typeorm.config';
+import dataSource from '../../../config/typeorm.config';
 import { Participant } from '../entities/participant.entity';
 
 export const participantController = {
